+++ example/_Test.Core/NLog.config
@@ -2,6 +2,7 @@
 <nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
   <extensions>
     <add assembly="NLog.MailKit" />
+    <add assembly="NLog.Targets.ElasticSearch"/>
   </extensions>
   <targets>
     <!--使用可自定义的着色将日志消息写入控制台-->
@@ -13,12 +14,19 @@
       <!--项目日志保存文件路径说明fileName="${basedir}/保存目录，以年月日的格式创建/${shortdate}/${记录器名称}-${单级记录}-${shortdate}.txt"-->
       <target name="log_file" xsi:type="File" fileName="${basedir}/Logs/${shortdate}/${logger}/${level}.txt" layout="[${level}] ${longdate} | ${message} ${onexception:${exception:format=message} ${newline} ${stacktrace} ${newline}" archiveFileName="${basedir}/archives/${logger}-${level}-${shortdate}-{#####}.txt" archiveAboveSize="102400" archiveNumbering="Sequence" concurrentWrites="true" keepFileOpen="false" />
     </target>
+    <target name="Elk" xsi:type="ElasticSearch" index="test" uri ="http://192.168.10.100:9200" requireAuth="false">
+      <field name="Level" layout="${level}" />
+      <field name="Host" layout="${machinename}" />
+      <field name="Time" layout="${longdate}" />
+      <field name="Message" layout="${message}" />
+    </target>
   </targets>
   <!--规则配置,final - 最终规则匹配后不处理任何规则-->
   <rules>
     <logger name="*" minlevel="Debug" writeTo="ColorConsole" />
-    <logger name="*" minlevel="Info" writeTo="Mail" />
+    <logger name="*" minlevel="Debug" writeTo="Elk" />
     <logger name="*" minlevel="Info" writeTo="AsyncFile" />
+    <logger name="*" minlevel="Error" writeTo="Mail" />
     <logger name="Microsoft.*" minlevel="Info" writeTo="" final="true" />
   </rules>
 </nlog>
