--- example/_Test.Core/a.json
@@ -1,6 +1,0 @@
-{
-  "A": {
-    "a1": 1,
-    "a2": "2"
-  }
-}
