--- example/_Test.Core/b.json
@@ -1,6 +1,0 @@
-{
-  "B": {
-    "b1": 3,
-    "b2": "4"
-  }
-}
