+++ src/Lib.Core/Plugins/Enviroment/EnviromentUtil.cs
@@ -1,0 +1,14 @@
+﻿namespace DwFramework.Core.Plugins
+{
+    public static class EnviromentUtil
+    {
+        /// <summary>
+        /// 获取系统信息
+        /// </summary>
+        /// <returns></returns>
+        public static SystemInfo GetSystemInfo()
+        {
+            return new SystemInfo();
+        }
+    }
+}
